Move EventColors out of event-details-boundary
