+++ lib/events/index.ts
@@ -1,2 +1,1 @@
-export * from "../../event-details-boundary/Event"
 export * from "./SaveEvent"
